Restre routes
Separated routes from contrallers
