+++ src/routes/book.ts
@@ -1,0 +1,15 @@
+import { Router } from 'express';
+import { addBook, deleteBook, getBookById, getBooks, updateBook } from '@entities/Book/Book.controller';
+
+const router = Router();
+
+
+router
+    .get('', getBooks)
+    .get('/:id', getBookById)
+    .post('', addBook)
+    .put('/:id', updateBook)
+    .delete('/:id', deleteBook);
+
+
+export default router;
