+++ src/routes/user.ts
@@ -1,0 +1,10 @@
+import { Router } from "express";
+import { getUsers, getUserById } from "../entities/User/User.controller"
+
+const router = Router();
+
+router
+    .get("", getUsers)
+    .get("/:id", getUserById);
+
+export default router;
